+++ domain/consensus/processes/transactionvalidator/transactionvalidator.go
@@ -18,7 +18,7 @@ type transactionValidator struct {
 	daaBlocksStore                          model.DAABlocksStore
 	enableNonNativeSubnetworks              bool
 	maxCoinbasePayloadLength                uint64
-	MergeSetSizeLimit                       uint
+	MergeSetSizeLimit                       uint64
 	coinbasePayloadScriptPublicKeyMaxLength uint8
 	sigCache                                *txscript.SigCache
 	sigCacheECDSA                           *txscript.SigCacheECDSA
@@ -29,7 +29,7 @@ type transactionValidator struct {
 func New(blockCoinbaseMaturity uint64,
 	enableNonNativeSubnetworks bool,
 	maxCoinbasePayloadLength uint64,
-	MergeSetSizeLimit uint,
+	MergeSetSizeLimit uint64,
 	coinbasePayloadScriptPublicKeyMaxLength uint8,
 	databaseContext model.DBReader,
 	pastMedianTimeManager model.PastMedianTimeManager,
